common: share a panic-safe stderr logger between both binaries
StderrLogger was duplicated verbatim in fast-nix-gc and
fast-nix-optimise, and used eprintln!, which panics when the stderr
write fails (closed pipe, full disk) and would abort a GC
mid-deletion. Move it into fast_nix_common and ignore write errors.

Also hoist the dangling 'pub mod nix_config' declaration to the top
of lib.rs next to the other modules.
